+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.0.15
+* Rename storage to local storage
+* Session storage
 ## 0.0.14
 * Simplyfy typings for react observable hook
 ## 0.0.13